% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{diabetes_x_train}
\alias{diabetes_x_train}
\title{Diabetes data (x training set)}
\description{
Diabetes data (x training set)
}
\details{
Data taken from the R package \code{lars} regarding diabetes patients and divided into
training set and test set by \href{Horseshoe prior with Stan}{https://ariddell.org/horseshoe-prior-with-stan.html}.
}

