% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_draws.R
\name{draw_beta}
\alias{draw_beta}
\title{Draw beta parameters}
\usage{
draw_beta(mu_n, Lambda_n_inv, sigma)
}
\arguments{
\item{mu_n}{The posterior beta expectation}

\item{Lambda_n_inv}{The inverse of the posterior prescision matrix.}

\item{sigma}{The sigma parameter}
}
\description{
Draw beta parameters
}
\details{
See the file derivations for details on how to draw the betas.
}

