% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_draws.R
\name{draw_lambda}
\alias{draw_lambda}
\title{Draw sigma}
\usage{
draw_lambda(lambda, beta, sigma, tau)
}
\arguments{
\item{lambda}{Previous lambda draw}

\item{beta}{Beta parameters}

\item{sigma}{Sigma parameter}

\item{tau}{Tau parameter}
}
\description{
Draw sigma
}
\details{
See the file derivations for details on how to draw sigma.
}

