% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_draws.R
\name{draw_sigma}
\alias{draw_sigma}
\title{Draw sigma}
\usage{
draw_sigma(a_n, b_0, yty, mu_n, Lambda_n)
}
\arguments{
\item{a_n}{posterior shape}

\item{b_0}{prior scale}

\item{yty}{y^t * y}

\item{mu_n}{The posterior beta expectation}
}
\description{
Draw sigma
}
\details{
See the file derivations for details on how to draw sigma.
}

