% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_draws.R
\name{draw_tau}
\alias{draw_tau}
\title{Draw tau}
\usage{
draw_tau(lambda, beta, sigma, tau)
}
\arguments{
\item{lambda}{Previous lambda draw}

\item{beta}{Beta parameters}

\item{sigma}{Sigma parameter}

\item{tau}{Tau parameter}

\item{shape_tau}{Tau shape parameter}
}
\description{
Draw tau
}
\details{
See the file derivations for details on how to draw tau.
}

