% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hslm.R, R/hsml-package.R
\docType{package}
\name{hslm}
\alias{hslm}
\alias{hslm-package}
\title{Bayeisan linear regression using the horseshoe prior}
\usage{
hslm(y, x, iter = 2000, intercept = FALSE, ab = c(1, 1))
}
\arguments{
\item{y}{dependent variabel as a column matrix.}

\item{x}{independent variables in a matrix.}

\item{iter}{The number of MCMC iterations. Default is 2000.}

\item{intercept}{Should an intercept be included in the model. Default is FALSE.}

\item{ab}{prior for sigma in model. If NULL, sigma is assumed to be known and 1. Default is c(1, 1).}
}
\value{
List with
\describe{
  \item{\code{beta}}{Matrix of size \code{iter} * (\code{ncol(x) + intercept})}
  \item{\code{lambda}}{Matrix of size \code{iter} * (\code{ncol(x) + intercept})}
  \item{\code{tau}}{Vector of length \code{iter}}
  \item{\code{sigma}}{Vector of length \code{iter}, if estimated.}
}
}
\description{
Function to estimate a linear regression with the horseshoe prior.
Model specification:
\deqn{y ~ N(X\beta^T,\sigma ^2)}
\deqn{\beta | \lambda_i, \tau ~ N(0, \tau^2  \lambda_i^2 \sigma^2)}
\deqn{\sigma^2 ~ IG(a, b)}
\deqn{\lambda_i ~ C_+(0, 1)}
\deqn{\tau ~ C_+(0, 1)}

Bayesian linear regression with the horseshoe prior.
}
\examples{
data(diabetes_x_train)
data(diabetes_y_train)
hs_res <- hslm(diabetes_y_train, diabetes_x_train)
}
\references{
\itemize{
   \item{Carvalho, C. M., Polson, N. G., & Scott, J. G. (2010). The horseshoe estimator for sparse signals. Biometrika}
}
}

